Move models to common module
