+++ aurora/common/models/battery_experiment.py
