+++ aurora/common/models/pattern_observer.py
