+++ aurora/interface/menus/submit_experiment.py
@@ -5,12 +5,12 @@ import ipywidgets as ipw
 from aiida_aurora.schemas.battery import BatterySample
 from aiida_aurora.schemas.utils import dict_to_formatted_json
 
+from aurora.common.models import AvailableSamplesModel, BatteryExperimentModel
 from aurora.engine import submit_experiment
 from aurora.interface.cycling import CyclingCustom, CyclingStandard
 from aurora.interface.sample import (SampleFromId, SampleFromRecipe,
                                      SampleFromSpecs)
 from aurora.interface.tomato import TomatoSettings
-from aurora.models import AvailableSamplesModel, BatteryExperimentModel
 
 CODE_NAME = "ketchup-0.2rc2"
